pipeline/airflow/requirements.txt
shap==0.46.0
numpy==1.26.4
fairlearn==0.11.0
fredapi==0.5.2
gcsfs==2024.10.0
google_api_python_client==1.12.11
joblib==1.4.2
matplotlib==3.9.2
omegaconf==2.3.0
pandas==2.2.3
protobuf==3.20.3
pytest==8.3.3
python-dotenv==1.0.1
PyYAML==6.0.2
Requests==2.32.3
scikit_learn==1.5.2
seaborn==0.13.2            
lightgbm==3.3.5           
torch==2.2.2
xgboost==2.1.2
yfinance==0.2.48
pymysql